requirements.txt
requests>=2.32.3
python-telegram-bot>=21.7